refactor(helpers): new folder structure for helpers and dev-tools #866
