+++ libs/angular-components/community/helpers/unique-id.ts
