+++ libs/angular-components/tedi/components/layout/grid/row/row.stories.ts
@@ -6,7 +6,7 @@ import {
 } from "@storybook/angular";
 import { RowComponent } from "./row.component";
 import { ColComponent } from "../col/col.component";
-import { createBreakpointArgTypes } from "../../../../../utils/createBreakpointArgTypes";
+import { createBreakpointArgTypes } from "../../../../../dev-tools/createBreakpointArgTypes";
 
 /**
  * <a href="https://tedi.tehik.ee/1ee8444b7/p/24da19-grid" target="_BLANK">Zeroheight ↗</a><br/>
